Make git_resources header to private
This change makes the git_resources header an internal header, i.e.
unavailable to library users.  The overall goal is to keep libgit2
headers from being visible to library users.
